+++ Source/RankingUp.Tournament.Application/Interfaces/IRankingGameService.cs
@@ -1,0 +1,14 @@
+﻿using RankingUp.Core.Domain;
+using RankingUp.Tournament.Application.ViewModels;
+
+namespace RankingUp.Tournament.Application.Interfaces
+{
+    public interface IRankingGameService
+    {
+        Task<RequestResponse<IEnumerable<RankingGameDetailViewModel>>> GetGames(Guid RankingId, bool? IsFinished);
+        Task<RequestResponse<RankingGameDetailViewModel>> CreateGameUsingQueue(Guid TournamentId, int UseId);
+        Task<RequestResponse<RankingGameDetailViewModel>> CreateGame(RankingCreateGameViewModel model);
+        Task<RequestResponse<RankingGameDetailViewModel>> UpdateGame(RankingGameDetailViewModel model);
+
+    }
+}
