+++ Source/RankingUp.Tournament.Application/Services/RankingAppService.cs
@@ -4,6 +4,7 @@ using RankingUp.Core.Communication.Mediator;
 using RankingUp.Core.Domain;
 using RankingUp.Core.ViewModels;
 using RankingUp.Player.Domain.IRepositories;
+using RankingUp.Tournament.Application.Interfaces;
 using RankingUp.Tournament.Application.ViewModels;
 using RankingUp.Tournament.Domain.Entities;
 using RankingUp.Tournament.Domain.Entities.Filters;
@@ -16,26 +17,20 @@ namespace RankingUp.Tournament.Application.Services
     public class RankingAppService : IRankingAppService
     {
         private readonly ITournamentsRepository _tournamentsRepository;
-        private readonly ITournamentTeamRepository _tournamentTeamRepository;
         private readonly ITournamentGameRepository _tournamentGameRepository;
-        private readonly IRankingQueueRepository _rankingQueueRepository;
         private readonly IClubRepository _clubRepository;
-        private readonly IPlayerRepository _playerRepository;
-        private readonly IPlayerClubsRepository _playerClubsRepository;
         private readonly IMapper _mapper;
         private readonly IMediatorHandler _mediatorHandler;
 
-        public RankingAppService(ITournamentsRepository tournamentsRepository, ITournamentTeamRepository tournamentTeamRepository
-            , ITournamentGameRepository tournamentGameRepository, IRankingQueueRepository rankingQueueRepository
-            , IClubRepository clubRepository, IPlayerRepository playerRepository, IPlayerClubsRepository playerClubsRepository, IMapper mapper, IMediatorHandler mediatorHandler)
+        public RankingAppService(ITournamentsRepository tournamentsRepository,
+            ITournamentGameRepository tournamentGameRepository,
+            IClubRepository clubRepository,
+            IMapper mapper,
+            IMediatorHandler mediatorHandler)
         {
             _tournamentsRepository = tournamentsRepository;
-            _tournamentTeamRepository = tournamentTeamRepository;
             _tournamentGameRepository = tournamentGameRepository;
-            _rankingQueueRepository = rankingQueueRepository;
             _clubRepository = clubRepository;
-            _playerRepository = playerRepository;
-            _playerClubsRepository = playerClubsRepository;
             _mapper = mapper;
             _mediatorHandler = mediatorHandler;
         }
@@ -83,24 +78,12 @@ namespace RankingUp.Tournament.Application.Services
                 return new RequestResponse<RankingDetailViewModel>(ex.Message);
             }
         }
-        public async Task<RequestResponse<IEnumerable<RankingPlayerViewModel>>> GetPlayers(Guid RankingId)
-        {
-            try
-            {
-                return new RequestResponse<IEnumerable<RankingPlayerViewModel>>(
-                    this._mapper.Map<IEnumerable<RankingPlayerViewModel>>(await _tournamentTeamRepository.GetAllByTournament(RankingId))
-                    , new Notifiable());
-            }
-            catch (Exception ex)
-            {
-                return new RequestResponse<IEnumerable<RankingPlayerViewModel>>(ex.Message);
-            }
-        }
+
         public async Task<RequestResponse<IEnumerable<RankingGameDetailViewModel>>> GetGames(Guid RankingId, bool IsFinished = false)
         {
             try
             {
-                if(!IsFinished)
+                if (!IsFinished)
                     return new RequestResponse<IEnumerable<RankingGameDetailViewModel>>(
                    this._mapper.Map<IEnumerable<RankingGameDetailViewModel>>(await _tournamentGameRepository.GetGameNotFinishTournamentId(RankingId))
                    , new Notifiable());
@@ -115,19 +98,7 @@ namespace RankingUp.Tournament.Application.Services
                 return new RequestResponse<IEnumerable<RankingGameDetailViewModel>>(ex.Message);
             }
         }
-        public async Task<RequestResponse<IEnumerable<RankingPlayerQueueViewModel>>> GetPlayersOnQueue(Guid RankingId)
-        {
-            try
-            {
-                return new RequestResponse<IEnumerable<RankingPlayerQueueViewModel>>(
-                    this._mapper.Map<IEnumerable<RankingPlayerQueueViewModel>>(await _rankingQueueRepository.GetByTournamentIdOrderByCreateDate(RankingId))
-                    , new Notifiable());
-            }
-            catch (Exception ex)
-            {
-                return new RequestResponse<IEnumerable<RankingPlayerQueueViewModel>>(ex.Message);
-            }
-        }
+
 
         public async Task<RequestResponse<IEnumerable<RankingDetailViewModel>>> GetRankingsByClub(Guid ClubId)
         {
@@ -187,7 +158,7 @@ namespace RankingUp.Tournament.Application.Services
                 if (orig is null)
                     throw new Exception("Ranking não encontrado!");
 
-                if(orig.IsFinish)
+                if (orig.IsFinish)
                     noticable.AddNotification("Ranking já foi finalizado!");
 
                 if (!orig.IsActive)
@@ -238,203 +209,6 @@ namespace RankingUp.Tournament.Application.Services
             return new NoContentResponse(noticable);
         }
 
-       
-
-        public async Task<RequestResponse<RankingPlayerViewModel>> AddPlayer(RankingPlayerViewModel model)
-        {
-            var noticable = new Notifiable();
-            try
-            {
-                var team = new TournamentTeam
-                    (await _tournamentsRepository.GetById(model.TournamentUUId),
-                    await _playerRepository.GetById(model.PlayerUUId), true, model.UserId);
-
-                team.Validate();
-                noticable.AddNotifications(team.Notifications);
-                if (team.Tournament?.IsFinish ?? false)
-                    noticable.AddNotification("O Ranking já foi finalizado");
-
-                if(team.Tournament != null && team.Tournament.OnlyClubMembers)
-                {
-                    var playersClub = await _playerClubsRepository.GetPlayerAndClubId(team.Tournament.ClubId, team.TeamId);
-                    if(playersClub is null)
-                        noticable.AddNotification("Esse Ranking somente permite jogadores associados ao clube");
-                }
-
-                var playerInRanking = await _tournamentTeamRepository.GetAllByTournament(model.TournamentUUId);
-                if(playerInRanking.Any(x=> x.Player.UUId == model.PlayerUUId))
-                    noticable.AddNotification("Jogador já cadastro no Ranking");
-
-
-                if (noticable.Valid)
-                {
-                    using (var scope = new TransactionScope(TransactionScopeAsyncFlowOption.Enabled))
-                    {
-                        team = await _tournamentTeamRepository.InsertAsync(team);
-                        scope.Complete();
-                    }
-                    await _mediatorHandler.PublishDomainEvent(new PlayerInRankingEvent(team.UUId,team.Tournament.UUId,model.UserId,Domain.Enums.PlayerRankingActionEnum.Added));
-                    return new RequestResponse<RankingPlayerViewModel>(_mapper.Map<RankingPlayerViewModel>(await _tournamentTeamRepository.GetById(team.Id)), noticable);
-                }
-            }
-            catch (Exception ex)
-            {
-                noticable.AddNotification(ex.Message);
-            }
-            return new RequestResponse<RankingPlayerViewModel>(noticable);
-        }
-        public async Task<RequestResponse<RankingGameDetailViewModel>> CreateGame(RankingCreateGameViewModel model)
-        {
-            var noticable = new Notifiable();
-            try
-            {
-                var game = new TournamentGame(
-                    (await _tournamentTeamRepository.GetById(model.TeamOneUUId)),
-                    (await _tournamentTeamRepository.GetById(model.TeamTwoUUId)),
-                    (await _tournamentsRepository.GetById(model.TournamentUUId)), model.UserId);
-
-                game.Validate();
-
-                noticable.AddNotifications(game.Notifications);
-
-                if (game.Tournament?.IsFinish ?? false)
-                    noticable.AddNotification("O Ranking já foi finalizado");
-
-                if(game.Tournament != null)
-                {
-                    var gamePlaying = await _tournamentGameRepository.GetAllGamesByTournamentId(game.Tournament.UUId);
-                    if(gamePlaying !=null && gamePlaying.Count() >= game.Tournament.MatchSameTime)
-                        noticable.AddNotification("O numero maximo de jogos ao mesmo tempo foi atingido");
-                }
-
-                if (noticable.Valid)
-                {
-                    using (var scope = new TransactionScope(TransactionScopeAsyncFlowOption.Enabled))
-                    {
-                        game = await _tournamentGameRepository.InsertAsync(game);
-                        scope.Complete();
-                    }
-                    await _mediatorHandler.PublishDomainEvent(new RankingGameCreatedEvent(game.UUId, game.Tournament.UUId, game.TeamOne.UUId, game.TeamTwo.UUId,game.CreatePersonId));
-                    return new RequestResponse<RankingGameDetailViewModel>(_mapper.Map<RankingGameDetailViewModel>(await _tournamentGameRepository.GetById(game.Id)), noticable);
-                }
-            }
-            catch (Exception ex)
-            {
-                noticable.AddNotification(ex.Message);
-            }
-            return new RequestResponse<RankingGameDetailViewModel>(noticable);
-        }
-        public async Task<RequestResponse<RankingGameDetailViewModel>> CreateGameUsingQueue(Guid TournamentId, int UseId)
-        {
-            var noticable = new Notifiable();
-            try
-            {
-                var playersInQueue = await this._rankingQueueRepository.GetByTournamentIdOrderByCreateDate(TournamentId);
-                if (!playersInQueue.Any() || playersInQueue.Count() <= 1)
-                    throw new Exception("Não há jogadores suficientes na Fila");
-
-                var games = await this._tournamentGameRepository.GetAllGamesByTournamentId(TournamentId);
-                var team1 = playersInQueue.First().Team;
-
-                TournamentTeam team2 = null;
-                int gamesPlayedWithTeam1 = int.MaxValue;
-                foreach (var rankingQueue in playersInQueue.Where(x => x.TeamId != team1.Id))
-                {
-                    var gamesPlayeds = games.Where(x => x.TeamOneId == rankingQueue.TeamId || x.TeamTwoId == rankingQueue.TeamId);
-                    if (!gamesPlayeds.Any())
-                    {
-                        gamesPlayedWithTeam1 = 0;
-                        team2 = rankingQueue.Team;
-                        break;
-                    }
-                    var gamesWithTeam1 = games.Where(x => x.TeamOneId == rankingQueue.TeamId || x.TeamTwoId == team1.Id).Count();
-                    gamesWithTeam1 += games.Where(x => x.TeamTwoId == rankingQueue.TeamId || x.TeamOneId == team1.Id).Count();
-                    
-                    if(gamesWithTeam1 < gamesPlayedWithTeam1)
-                    {
-                        gamesPlayedWithTeam1 = gamesWithTeam1;
-                        team2 = rankingQueue.Team;
-                    }
-                }
-
-                var game = new TournamentGame(
-                    team1,
-                    team2,
-                    (await _tournamentsRepository.GetById(TournamentId)), UseId);
-
-                game.Validate();
-                noticable.AddNotifications(game.Notifications);
-
-                if (game.Tournament?.IsFinish ?? false)
-                    noticable.AddNotification("O Ranking já foi finalizado");
-
-                if (game.Tournament != null)
-                {
-                    var gamePlaying = await _tournamentGameRepository.GetAllGamesByTournamentId(game.Tournament.UUId);
-                    if (gamePlaying != null && gamePlaying.Count() >= game.Tournament.MatchSameTime)
-                        noticable.AddNotification("O numero maximo de jogos ao mesmo tempo foi atingido");
-                }
-
-                if (noticable.Valid)
-                {
-                    using (var scope = new TransactionScope(TransactionScopeAsyncFlowOption.Enabled))
-                    {
-                        game = await _tournamentGameRepository.InsertAsync(game);
-                        scope.Complete();
-                    }
-                    await _mediatorHandler.PublishDomainEvent(new RankingGameCreatedEvent(game.UUId, game.Tournament.UUId, game.TeamOne.UUId, game.TeamTwo.UUId, game.CreatePersonId));
-                    return new RequestResponse<RankingGameDetailViewModel>(_mapper.Map<RankingGameDetailViewModel>(await _tournamentGameRepository.GetById(game.Id)), noticable);
-                }
-            }
-            catch (Exception ex)
-            {
-                noticable.AddNotification(ex.Message);
-            }
-            return new RequestResponse<RankingGameDetailViewModel>(noticable);
-        }
-
-
-
-        public async Task<RequestResponse<RankingGameDetailViewModel>> UpdateGame(RankingGameDetailViewModel model)
-        {
-            var noticable = new Notifiable();
-            try
-            {
-                var orig = await _tournamentGameRepository.GetById(model.UUId);
-                if (orig is null)
-                    throw new Exception("Jogo não encontrado!");
-
-                var game = new TournamentGame(model.TeamOneGamePoints,model.TeamTwoGamePoints,
-                    orig.TeamOne,
-                    orig.TeamTwo,
-                    (await _tournamentsRepository.GetById(orig.TournamentId)),
-                    model.UserId, model.IsFinished);
-
-                game.Validate();
-                noticable.AddNotifications(game.Notifications);
-
-                if (game.Tournament?.IsFinish ?? false)
-                    noticable.AddNotification("O Ranking já foi finalizado");
-
-                if (noticable.Valid)
-                {
-                    game.Id = orig.Id;
-                    game.UUId = orig.UUId;
-                    using (var scope = new TransactionScope(TransactionScopeAsyncFlowOption.Enabled))
-                    {
-                        await _tournamentGameRepository.UpdateAsync(game);
-                        scope.Complete();
-                    }
-                    await _mediatorHandler.PublishDomainEvent(new RankingGameUpdatedEvent(game.UUId, game.Tournament.UUId, game.IsFinished, game.TeamOne.UUId, game.TeamTwo.UUId, game.UpdatePersonId));
-                    return new RequestResponse<RankingGameDetailViewModel>(_mapper.Map<RankingGameDetailViewModel>(await _tournamentGameRepository.GetById(game.Id)), noticable);
-                }
-            }
-            catch (Exception ex)
-            {
-                noticable.AddNotification(ex.Message);
-            }
-            return new RequestResponse<RankingGameDetailViewModel>(noticable);
-        }
         public async Task<RequestResponse<RankingDetailViewModel>> UpdateRanking(RankingDetailViewModel model)
         {
             var noticable = new Notifiable();
@@ -501,38 +275,8 @@ namespace RankingUp.Tournament.Application.Services
             }
             return new NoContentResponse(noticable);
         }
-        public async Task<NoContentResponse> RemovePlayer(Guid Id, int UseId)
-        {
-            var noticable = new Notifiable();
-            try
-            {
-                var orig = await _tournamentTeamRepository.GetById(Id);
 
-                if (orig is null)
-                    throw new Exception("Jogador não encontrado!");
 
-                var playerIsPlaying = await _tournamentGameRepository.GetGameNotFinishByTeamAndTournamentId(orig.Tournament.UUId, orig.UUId);
-                if (playerIsPlaying != null)
-                    noticable.AddNotification("Não pode remover um jogador em uma partida em andamento!");
 
-                orig.Disable(UseId);
-                if (noticable.Valid)
-                {
-                    using (var scope = new TransactionScope(TransactionScopeAsyncFlowOption.Enabled))
-                    {
-                        await _tournamentTeamRepository.DeleteAsync(orig);
-                        scope.Complete();
-                    }
-                    await _mediatorHandler.PublishDomainEvent(new PlayerInRankingEvent(orig.UUId, orig.Tournament.UUId, UseId, Domain.Enums.PlayerRankingActionEnum.Deleted));
-                }
-            }
-            catch (Exception ex)
-            {
-                noticable.AddNotification(ex.Message);
-            }
-            return new NoContentResponse(noticable);
-        }
-
-       
     }
 }
