+++ Source/RankingUp.WebApp.API/Program.cs
@@ -12,7 +12,7 @@ builder.Services.AddSignalR();
 builder.Services.RegisterServices(builder.Configuration);
 builder.Services.AddCors(policyBuilder =>
     policyBuilder.AddDefaultPolicy(policy =>
-        policy.WithOrigins("*").AllowAnyHeader().AllowAnyHeader())
+        policy.WithOrigins("*").AllowAnyMethod().AllowAnyHeader())
 );
 
 var app = builder.Build();
